+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # JVM Common Buildpack Changelog
 
-## master
+## v60
 
+* Added LD_LIBRARY_PATH to export and profile.d scripts 
 * Install heroku-java-metrics-agent with all apps
 * Convert .profile.d script to work with Dash and Bash